[DFS-20] Restructure repo and add initial Terraform files
